Change to custom Error type.
Some HALs implement the digital interface in a way that it cannot fail (e.g.
STM32). In this case Error type of Input and Output pins are Infallible.
With a custom Error type we can allow the use of `.into_ok()` when using such
HAL implementations.

Further this custom error type removes the type bound of `core::fmt::Debug` from
the Error type. This might be useful at some point.